backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
transformers==4.50.3
torch==2.1.2
pandas==2.1.1
scikit-learn==1.2.2
numpy==1.26.4
tqdm==4.62.3
pdfplumber==0.11.6
python-docx==1.1.2